refactor: move height/visibility update handlers into results component
